@@ -4,11 +4,14 @@
  */
 package cz.miroslavbartyzal.psdiagram.app.gui;
 
+import cz.miroslavbartyzal.psdiagram.app.global.MyExceptionHandler;
+import cz.miroslavbartyzal.psdiagram.app.global.SettingsHolder;
 import java.awt.Desktop;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import java.io.IOException;
 import java.net.URI;
+import javax.swing.JButton;
 
 /**
  * Tato třída představuje formulář o aplikaci.
@@ -18,7 +21,9 @@ import java.net.URI;
 public class JFrameAbout extends javax.swing.JFrame
 {
 
-    final URI uri = URI.create("http://www.psdiagram.cz");
+    final URI webUri = URI.create(SettingsHolder.PSDIAGRAM_SERVER);
+    final URI githubUri = URI.create("https://github.com/PoloShock/psdiagram");
+    final URI facebookUri = URI.create("http://facebook.com/psdiagram");
 
     // TODO email dat jako selectable
     /**
@@ -29,24 +34,9 @@ public class JFrameAbout extends javax.swing.JFrame
 
         initComponents();
 
-//        jButtonWeb.setBackground(new Color(0, 0, 0, 0));
-        jButtonWeb.setToolTipText(uri.toString());
-        jButtonWeb.addActionListener(new ActionListener()
-        {
-            @Override
-            public void actionPerformed(ActionEvent e)
-            {
-                if (Desktop.isDesktopSupported()) {
-                    try {
-                        Desktop.getDesktop().browse(uri);
-                    } catch (IOException ex) {
-                        ex.printStackTrace(System.err);
-                    }
-                } else {
-                    // TODO: error handling
-                }
-            }
-        });
+        initUriButton(jButtonWeb, webUri);
+        initUriButton(jButtonGitWeb, githubUri);
+        initUriButton(jButtonFbWeb, facebookUri);
 
         jLabel6.setText(
                 jLabel6.getText() + cz.miroslavbartyzal.psdiagram.app.global.SettingsHolder.PSDIAGRAM_VERSION);
@@ -56,6 +46,27 @@ public class JFrameAbout extends javax.swing.JFrame
 //        jPanel2.revalidate();
 //        jPanel1.revalidate();
 //        super.revalidate();
+    }
+    
+    private void initUriButton(JButton button, URI uri) {
+//        button.setBackground(new Color(0, 0, 0, 0));
+        button.setToolTipText(uri.toString());
+        button.addActionListener(new ActionListener()
+        {
+            @Override
+            public void actionPerformed(ActionEvent e)
+            {
+                if (Desktop.isDesktopSupported()) {
+                    try {
+                        Desktop.getDesktop().browse(uri);
+                    } catch (IOException ex) {
+                        MyExceptionHandler.handle(ex);
+                    }
+                } else {
+                    // TODO: error handling
+                }
+            }
+        });
     }
 
     /**
@@ -70,47 +81,130 @@ public class JFrameAbout extends javax.swing.JFrame
     {
 
         jPanel1 = new javax.swing.JPanel();
-        jLabelTitle = new javax.swing.JLabel();
         jLabel1 = new javax.swing.JLabel();
+        jLabelTitle = new javax.swing.JLabel();
+        jPanel3 = new javax.swing.JPanel();
+        jPanel4 = new javax.swing.JPanel();
+        jLabel5 = new javax.swing.JLabel();
+        jLabel8 = new javax.swing.JLabel();
+        jLabel7 = new javax.swing.JLabel();
         jLabel2 = new javax.swing.JLabel();
+        jButtonWeb = new javax.swing.JButton();
+        jLabel6 = new javax.swing.JLabel();
         jLabel3 = new javax.swing.JLabel();
         jLabel4 = new javax.swing.JLabel();
-        jLabel6 = new javax.swing.JLabel();
-        jLabel7 = new javax.swing.JLabel();
+        jButtonGitWeb = new javax.swing.JButton();
+        jButtonFbWeb = new javax.swing.JButton();
         jPanel2 = new javax.swing.JPanel();
         jTextPane1 = new javax.swing.JTextPane();
-        jButtonWeb = new javax.swing.JButton();
 
         setTitle("O apikaci PS Diagram");
         setAlwaysOnTop(true);
-        setResizable(false);
         setType(java.awt.Window.Type.UTILITY);
 
         jPanel1.setBorder(javax.swing.BorderFactory.createTitledBorder(""));
+
+        jLabel1.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        jLabel1.setIcon(new javax.swing.ImageIcon(getClass().getResource("/img/about.png"))); // NOI18N
 
         jLabelTitle.setFont(new java.awt.Font("sansserif", 1, 24)); // NOI18N
         jLabelTitle.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
         jLabelTitle.setText("PS Diagram");
 
-        jLabel1.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        jLabel1.setIcon(new javax.swing.ImageIcon(getClass().getResource("/img/about.png"))); // NOI18N
+        jLabel5.setText("GitHub: ");
+
+        jLabel8.setText("Facebook: ");
+
+        jLabel7.setText("Build: ");
 
         jLabel2.setText("Autor: Miroslav Bartyzal");
+
+        jButtonWeb.setText("<html>\n<FONT color=\\\"#000099\\\"><U>http://psdiagram.cz</U></FONT>\n</html>");
+        jButtonWeb.setAlignmentY(0.0F);
+        jButtonWeb.setBorder(null);
+        jButtonWeb.setBorderPainted(false);
+        jButtonWeb.setContentAreaFilled(false);
+        jButtonWeb.setCursor(new java.awt.Cursor(java.awt.Cursor.HAND_CURSOR));
+
+        jLabel6.setText("Verze: ");
 
         jLabel3.setText("Email: miroslavbartyzal@gmail.com");
 
         jLabel4.setText("Web: ");
 
-        jLabel6.setText("Verze: ");
+        jButtonGitWeb.setText("<html>\n<FONT color=\\\"#000099\\\"><U>https://github.com/PoloShock/psdiagram</U></FONT>\n</html>");
+        jButtonGitWeb.setAlignmentY(0.0F);
+        jButtonGitWeb.setBorder(null);
+        jButtonGitWeb.setBorderPainted(false);
+        jButtonGitWeb.setContentAreaFilled(false);
+        jButtonGitWeb.setCursor(new java.awt.Cursor(java.awt.Cursor.HAND_CURSOR));
 
-        jLabel7.setText("Build: ");
+        jButtonFbWeb.setText("<html>\n<FONT color=\\\"#000099\\\"><U>http://facebook.com/psdiagram</U></FONT>\n</html>");
+        jButtonFbWeb.setAlignmentY(0.0F);
+        jButtonFbWeb.setBorder(null);
+        jButtonFbWeb.setBorderPainted(false);
+        jButtonFbWeb.setContentAreaFilled(false);
+        jButtonFbWeb.setCursor(new java.awt.Cursor(java.awt.Cursor.HAND_CURSOR));
+
+        javax.swing.GroupLayout jPanel4Layout = new javax.swing.GroupLayout(jPanel4);
+        jPanel4.setLayout(jPanel4Layout);
+        jPanel4Layout.setHorizontalGroup(
+            jPanel4Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(jPanel4Layout.createSequentialGroup()
+                .addContainerGap()
+                .addGroup(jPanel4Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addComponent(jLabel7)
+                    .addComponent(jLabel6)
+                    .addGroup(jPanel4Layout.createSequentialGroup()
+                        .addComponent(jLabel4)
+                        .addGap(0, 0, 0)
+                        .addComponent(jButtonWeb, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
+                    .addGroup(jPanel4Layout.createSequentialGroup()
+                        .addComponent(jLabel5)
+                        .addGap(0, 0, 0)
+                        .addComponent(jButtonGitWeb, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
+                    .addGroup(jPanel4Layout.createSequentialGroup()
+                        .addComponent(jLabel8)
+                        .addGap(0, 0, 0)
+                        .addComponent(jButtonFbWeb, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
+                    .addComponent(jLabel3)
+                    .addComponent(jLabel2))
+                .addContainerGap())
+        );
+        jPanel4Layout.setVerticalGroup(
+            jPanel4Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(jPanel4Layout.createSequentialGroup()
+                .addContainerGap()
+                .addComponent(jLabel2)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(jLabel3)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addGroup(jPanel4Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                    .addComponent(jLabel4)
+                    .addComponent(jButtonWeb, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addGroup(jPanel4Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                    .addComponent(jLabel5)
+                    .addComponent(jButtonGitWeb, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addGroup(jPanel4Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
+                    .addComponent(jLabel8)
+                    .addComponent(jButtonFbWeb, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(jLabel6)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(jLabel7)
+                .addContainerGap())
+        );
+
+        jPanel3.add(jPanel4);
 
         jPanel2.setBorder(javax.swing.BorderFactory.createTitledBorder("Další informace"));
 
         jTextPane1.setEditable(false);
         jTextPane1.setBorder(null);
         jTextPane1.setContentType("text/html"); // NOI18N
-        jTextPane1.setText("<html>\n<head>\n<style type=\"text/css\">\nbody { background-color:D6D9DF; }\n</style>\n</head>\n<body>\nAplikace vznikla jako bakalářská práce v roce 2011-2012.<br /><br />\nMáte-li jakékoliv dotazy, připomínky nebo nápady na vylepšení, neváhejte mne prosím kontaktovat na výše uvedeném emailu.\n</body>\n</html>");
+        jTextPane1.setText("<html>\n<head>\n<style type=\"text/css\">\nbody { background-color:D6D9DF; }\n</style>\n</head>\n<body>\nAplikace vznikla jako bakalářská práce v roce 2011-2012.<br /><br />\nMáte-li jakékoliv dotazy, připomínky nebo nápady na vylepšení, neváhejte mne prosím kontaktovat na výše uvedeném emailu, GitHubu nebo Facebooku.\n</body>\n</html>");
 
         javax.swing.GroupLayout jPanel2Layout = new javax.swing.GroupLayout(jPanel2);
         jPanel2.setLayout(jPanel2Layout);
@@ -123,32 +217,14 @@ public class JFrameAbout extends javax.swing.JFrame
             .addComponent(jTextPane1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
         );
 
-        jButtonWeb.setText("<html>\n<FONT color=\\\"#000099\\\"><U>www.psdiagram.cz</U></FONT>\n</html>");
-        jButtonWeb.setAlignmentY(0.0F);
-        jButtonWeb.setBorder(null);
-        jButtonWeb.setBorderPainted(false);
-        jButtonWeb.setContentAreaFilled(false);
-        jButtonWeb.setCursor(new java.awt.Cursor(java.awt.Cursor.HAND_CURSOR));
-
         javax.swing.GroupLayout jPanel1Layout = new javax.swing.GroupLayout(jPanel1);
         jPanel1.setLayout(jPanel1Layout);
         jPanel1Layout.setHorizontalGroup(
             jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addComponent(jLabel1, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+            .addComponent(jLabel1, javax.swing.GroupLayout.DEFAULT_SIZE, 378, Short.MAX_VALUE)
             .addComponent(jLabelTitle, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
             .addComponent(jPanel2, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-            .addGroup(jPanel1Layout.createSequentialGroup()
-                .addGap(89, 89, 89)
-                .addGroup(jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-                    .addComponent(jLabel7)
-                    .addComponent(jLabel6)
-                    .addGroup(jPanel1Layout.createSequentialGroup()
-                        .addComponent(jLabel4)
-                        .addGap(0, 0, 0)
-                        .addComponent(jButtonWeb, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
-                    .addComponent(jLabel3)
-                    .addComponent(jLabel2))
-                .addContainerGap(85, Short.MAX_VALUE))
+            .addComponent(jPanel3, javax.swing.GroupLayout.Alignment.TRAILING, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
         );
         jPanel1Layout.setVerticalGroup(
             jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
@@ -157,17 +233,7 @@ public class JFrameAbout extends javax.swing.JFrame
                 .addGap(0, 0, 0)
                 .addComponent(jLabelTitle)
                 .addGap(18, 18, 18)
-                .addComponent(jLabel2)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addComponent(jLabel3)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addGroup(jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.BASELINE)
-                    .addComponent(jLabel4)
-                    .addComponent(jButtonWeb, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addComponent(jLabel6)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addComponent(jLabel7)
+                .addComponent(jPanel3, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
                 .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
                 .addComponent(jPanel2, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE))
         );
@@ -183,22 +249,28 @@ public class JFrameAbout extends javax.swing.JFrame
             .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, layout.createSequentialGroup()
                 .addGap(1, 1, 1)
                 .addComponent(jPanel1, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addContainerGap(15, Short.MAX_VALUE))
+                .addGap(31, 31, 31))
         );
 
         pack();
     }// </editor-fold>//GEN-END:initComponents
     // Variables declaration - do not modify//GEN-BEGIN:variables
+    private javax.swing.JButton jButtonFbWeb;
+    private javax.swing.JButton jButtonGitWeb;
     private javax.swing.JButton jButtonWeb;
     private javax.swing.JLabel jLabel1;
     private javax.swing.JLabel jLabel2;
     private javax.swing.JLabel jLabel3;
     private javax.swing.JLabel jLabel4;
+    private javax.swing.JLabel jLabel5;
     private javax.swing.JLabel jLabel6;
     private javax.swing.JLabel jLabel7;
+    private javax.swing.JLabel jLabel8;
     private javax.swing.JLabel jLabelTitle;
     private javax.swing.JPanel jPanel1;
     private javax.swing.JPanel jPanel2;
+    private javax.swing.JPanel jPanel3;
+    private javax.swing.JPanel jPanel4;
     private javax.swing.JTextPane jTextPane1;
     // End of variables declaration//GEN-END:variables
 
